\name{phenoData}
\alias{phenoData}
\docType{data}

\title{
Phenotype Data for the Lung Data
}

\description{
phenotype data related to the \code{lungData}.
}

\usage{data(phenoData)}

\format{
The data set has several components:
\describe{
\item{\code{ventilator}}{Data frame of the phenotype trait variable of ventilator. Indicator of if a ventilator is used immediately before death. 1 = ventilator used, 0 = otherwise.}
\item{\code{gender}}{Data frame of the phenotype trait variable of gender. 1 = male, 0 = female.}
}
}

\seealso{
\code{\link{lungData}}.
}

\source{
The Genotype Tissue Expression (GTEx) project: \url{http://www.gtexportal.org/home/}
}

%\examples{
%
%}

\keyword{datasets}


